Added Guestbook
It runs on an external page. The only downside is that I couldn't
remove the ink to these guys. It is being generated on load by the
javascript… I tried to write a little JQuery script to remove the
element but it has one of those nasty target="_blank" attributes for
the anchor